feat: move built `emoji-mart` package to root
As this git repository consists of [yarn workspaces](https://yarnpkg.com/features/workspaces),
which enable to define mono repo packages within one repository, but we want
to use only one package `emoji-mart`. `slido-frontend` does not use yarn2 yet, so the workspace
can't be specified see yarnpkg/yarn#1570 (comment).

As we are of the full control of this forked repository, the fastest and easiest is to change the
folder structure to consist only of this package 🎉.

I was also considering this 3rd party subfolder generation https://github.com/EqualMa/gitpkg, but decided
not to depend on potentially security risk of 3rd party provider.

The dist related properties were adjusted in the root package.json from  emoji-mart package, as
this root package is used as a reference when usign this dependency.