fix: don't tab through all tablist items
Tab should only focus the active item in the tablist, second Tab should
move the focus to the next element in the document.

"When focus moves into the tab list, places focus on the active tab element.
When the tab list contains the focus, moves focus to the next element in
the page tab sequence outside the tablist, which is the tabpanel unless
the first element containing meaningful content inside the tabpanel is focusable."

https://www.w3.org/WAI/ARIA/apg/patterns/tabs/

Relates to PS-20252.