feat: add kbd navigation into emojis list
This kbd navigation worked but surprisingly, one had to have focus in
the search field for it to work. Having focus in the emoji list and
pressing arrow keys however didn't work.

So I just took the search arrow key navigation and used the same for the
emojis list. It's kinda weird that I am using the input field when
calling navigate() but I didn't want to make too many changes and it
works like that just fine.

Relates to PS-20252.