fix: add missing keyboard navigation for categories
Previously, the category tabs were all tabbable. This is wrong according
to WAI-ARIA guidelines. I changed that to only allowing the current tab
to be tabbable, but broke the keyboard navigation by doing so since the
rest of the tabs wasn't tabbable anymore and the arrow keys weren't
working natively.

So I added a manual arrow key navigation. The tricky part was to handle
the fact that the full emoji list is always only partially rendered.
Practically, I can select an emoji in the 1st category, then select
let's say a 5th category in tablist, and now the 1st is not rendered
anymore. Which also means there would be no tabbable emoji for the user
to navigate to! For this reason, and for convenience as well, since user
should be able to tab to emojis of the just selected category, I always
make the first emoji in the currently selected category tabbable upon tab
selection.

Relates to PS-21279.