fix: make emojis navigation more accessible in NVDA
In the previous commits, I accepted the fact that NVDA user has to switch
from Browse mode to Focus mode to navigate through emojis. This was true
for our markup back then. But then I noticed, that user can use arrow
keys in the categories tablist just fine in the Browse mode. So I wondered
if I could make the emoji list more accessible by assigning the right
roles. And I found that yes, doing so allows the user to navigate through
emojis using arrow keys directly in the Browse mode!

I was torn between using `radiogroup` and `listbox`. The 2 are similar
in many ways, with the exception that only 1 radio can be selected at a
time whereas list box can have more. This is however not important
because only 1 emoji can be selected at a time. In the end, I went with
listbox because it sounds like a more natural choice for a list of items
but I don't have a strong opinion there.

Closes PS-21279.